requirements.txt
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
facebook-sdk==3.1.0
Flask==1.0.3
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.0
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.0.0
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pysrt==1.1.1
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
